render deÃlete movie button only for logged users with admin role

+++ src/components/moviesTable.js
@@ -1,7 +1,10 @@
 import React from 'react';
+import auth from '../services/authService';
 import Like from './common/like';
 import Table from './common/table';
 import { Link } from 'react-router-dom';
+
+const user = auth.getCurrentUser();
 
 const MoviesTable = ({ movies, sortColumn, onDelete, onLike, onSort }) => {
   const columns = [
@@ -18,20 +21,25 @@ const MoviesTable = ({ movies, sortColumn, onDelete, onLike, onSort }) => {
       content: movie => (
         <Like liked={movie.liked} onClick={() => onLike(movie)}></Like>
       )
-    },
-    {
-      key: 'delete',
-      content: movie => (
-        <button
-          onClick={() => onDelete(movie)}
-          id={movie._id}
-          className='btn btn-danger btn-sm'
-        >
-          Delete
-        </button>
-      )
     }
   ];
+
+  if (user && user.isAdmin) {
+    columns.push({
+      key: 'delete',
+      content: movie => {
+        return (
+          <button
+            onClick={() => onDelete(movie)}
+            id={movie._id}
+            className='btn btn-danger btn-sm'
+          >
+            Delete
+          </button>
+        );
+      }
+    });
+  }
 
   return (
     <Table
